refactor(webhooks): fix CodeQL and React Doctor findings, reorganize helpers
Compare parsed hostnames in the transport test instead of URL prefixes,
split the delivery details sheet into summary, attempt list and payload
components, and move helpers into their own files: crypto and DNS utils,
queue name constants, API error responses and runtime, dashboard
authorization, and shared zod input schemas.
