+++ apps/api/src/constants/webhooks.ts
@@ -1,0 +1,10 @@
+import { errorResponse } from "../utils/openapi-responses";
+
+export const WEBHOOK_ERROR_RESPONSES = {
+  400: errorResponse("Invalid webhook request"),
+  401: errorResponse("Unauthorized"),
+  402: errorResponse("Subscription required"),
+  403: errorResponse("Forbidden"),
+  404: errorResponse("Webhook not found"),
+  503: errorResponse("Webhooks unavailable"),
+};
