+++ apps/api/src/runtime/webhooks.ts
@@ -1,0 +1,4 @@
+import { postgresDatabaseLayer } from "@notra/webhooks/runtime/postgres";
+import { ManagedRuntime } from "effect";
+
+export const webhookRuntime = ManagedRuntime.make(postgresDatabaseLayer);
