+++ packages/webhooks/src/constants/queues.ts
@@ -1,0 +1,2 @@
+export const EVENT_QUEUE_NAME = "notra-webhook-events";
+export const DELIVERY_QUEUE_NAME = "notra-webhook-deliveries";
