+++ packages/webhooks/src/utils/hmac.ts
@@ -1,29 +1,25 @@
 import { Effect } from "effect";
 
 import { WebhookCryptoError } from "../errors/webhooks";
-import { base64, bytes } from "./encoding";
+import { cryptoOperation, decodeBase64 } from "./crypto";
+import { base64 } from "./encoding";
 
 const importSigningKey = Effect.fn("webhooks.importSigningKey")(function* (
   secret: string
 ) {
-  const raw = yield* Effect.try({
-    try: () => bytes(secret.replace(/^whsec_/, "")),
-    catch: () => new WebhookCryptoError({ operation: "hmac.decode" }),
-  });
+  const raw = yield* decodeBase64(secret.replace(/^whsec_/, ""), "hmac.decode");
   if (raw.length !== 32) {
     return yield* new WebhookCryptoError({ operation: "hmac.invalidSecret" });
   }
-  return yield* Effect.tryPromise({
-    try: () =>
-      crypto.subtle.importKey(
-        "raw",
-        raw,
-        { name: "HMAC", hash: "SHA-256" },
-        false,
-        ["sign", "verify"]
-      ),
-    catch: () => new WebhookCryptoError({ operation: "hmac.import" }),
-  });
+  return yield* cryptoOperation("hmac.import", () =>
+    crypto.subtle.importKey(
+      "raw",
+      raw,
+      { name: "HMAC", hash: "SHA-256" },
+      false,
+      ["sign", "verify"]
+    )
+  );
 });
 
 export const hmacSign = Effect.fn("webhooks.hmacSign")(function* (
@@ -31,11 +27,9 @@ export const hmacSign = Effect.fn("webhooks.hmacSign")(function* (
   message: string
 ) {
   const key = yield* importSigningKey(secret);
-  const signature = yield* Effect.tryPromise({
-    try: () =>
-      crypto.subtle.sign("HMAC", key, new TextEncoder().encode(message)),
-    catch: () => new WebhookCryptoError({ operation: "hmac.sign" }),
-  });
+  const signature = yield* cryptoOperation("hmac.sign", () =>
+    crypto.subtle.sign("HMAC", key, new TextEncoder().encode(message))
+  );
   return base64(new Uint8Array(signature));
 });
 
@@ -45,13 +39,8 @@ export const hmacVerify = Effect.fn("webhooks.hmacVerify")(function* (
   signature: string
 ) {
   const key = yield* importSigningKey(secret);
-  const raw = yield* Effect.try({
-    try: () => bytes(signature),
-    catch: () => new WebhookCryptoError({ operation: "hmac.decodeSignature" }),
-  });
-  return yield* Effect.tryPromise({
-    try: () =>
-      crypto.subtle.verify("HMAC", key, raw, new TextEncoder().encode(message)),
-    catch: () => new WebhookCryptoError({ operation: "hmac.verify" }),
-  });
+  const raw = yield* decodeBase64(signature, "hmac.decodeSignature");
+  return yield* cryptoOperation("hmac.verify", () =>
+    crypto.subtle.verify("HMAC", key, raw, new TextEncoder().encode(message))
+  );
 });
